Safari-609.1.11.1/LayoutTests/media/session/play-pause-media-events-in-non-content-sessions-expected.txt

'Transient', 'Transient Solo', and 'Ambient' sessions should not respond to Play/Pause media control events.


---
Testing "transient" session kind
---
EVENT(canplaythrough)
EVENT(playing)
Media began playing.
EXPECTED (video.paused == 'false') OK
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
Non-Content Media Sessions should not respond to Play/Pause events.
Sending Play/Pause media event.
RUN(internals.sendMediaControlEvent("play-pause"))
EXPECTED (video.paused == 'false') OK
---
Testing "transient-solo" session kind
---
EVENT(canplaythrough)
EVENT(playing)
Media began playing.
EXPECTED (video.paused == 'false') OK
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
Non-Content Media Sessions should not respond to Play/Pause events.
Sending Play/Pause media event.
RUN(internals.sendMediaControlEvent("play-pause"))
EXPECTED (video.paused == 'false') OK
---
Testing "ambient" session kind
---
EVENT(canplaythrough)
EVENT(playing)
Media began playing.
EXPECTED (video.paused == 'false') OK
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
Non-Content Media Sessions should not respond to Play/Pause events.
Sending Play/Pause media event.
RUN(internals.sendMediaControlEvent("play-pause"))
EXPECTED (video.paused == 'false') OK
END OF TEST
